+++ web/src/info_overlay.js
@@ -266,7 +266,7 @@ export function set_up_toggler() {
             $(".overlay-modal").hide();
             $(`#${CSS.escape(key)}`).show();
             scroll_util
-                .get_scroll_element($(`#${CSS.escape(key)}`).find(".modal-body"))
+                .get_scroll_element($(`#${CSS.escape(key)}`).find(".overlay-scroll-container"))
                 .trigger("focus");
         },
     };
@@ -277,7 +277,7 @@ export function set_up_toggler() {
 
     const modals = opts.values.map((item) => {
         const key = item.key; // e.g. message-formatting
-        const $modal = $(`#${CSS.escape(key)}`).find(".modal-body");
+        const $modal = $(`#${CSS.escape(key)}`).find(".overlay-scroll-container");
         return $modal;
     });