+++ web/src/recent_view_data.ts
@@ -12,6 +12,11 @@ export const topics = new Map<string, TopicData>();
 // For pms, key is the user IDs to whom the message is being sent.
 
 export function process_message(msg: Message): boolean {
+    // Important: This function must correctly handle processing a
+    // given message more than once; this happens during the loading
+    // process because of how recent_view_message_list_data duplicates
+    // all_messages_data.
+
     // Return whether any conversation data is updated.
     let conversation_data_updated = false;