+++ zerver/lib/url_redirects.py
@@ -77,6 +77,7 @@ HELP_DOCUMENTATION_REDIRECTS: List[URLRedirect] = [
     URLRedirect("/help/starting-a-new-private-thread", "/help/starting-a-new-direct-message"),
     URLRedirect("/help/edit-or-delete-a-message", "/help/delete-a-message"),
     URLRedirect("/help/start-a-new-topic", "/help/starting-a-new-topic"),
+    URLRedirect("/help/configure-default-view", "/help/configure-home-view"),
 ]
 
 LANDING_PAGE_REDIRECTS = [